+++ context/app/static/js/components/ServiceStatusTable/StatusIcon.jsx
