+++ context/app/static/js/components/ServiceStatusTable/index.js
@@ -1,0 +1,3 @@
+import ServiceStatusTable from './ServiceStatusTable';
+
+export default ServiceStatusTable;
